created page for order edition
