+++ backend/src/auth/guards/access-token.guard.ts
@@ -1,5 +1,15 @@
-import { Injectable } from "@nestjs/common";
-import { AuthGuard } from "@nestjs/passport";
+import { ExecutionContext, Injectable } from '@nestjs/common';
+
+import { AuthGuard } from '@nestjs/passport';
+import { Observable } from 'rxjs';
 
 @Injectable()
-export class AccessTokenGuard extends AuthGuard('jwt') {}
+export class AccessTokenGuard extends AuthGuard('jwt') {
+  canActivate(
+    context: ExecutionContext,
+  ): boolean | Promise<boolean> | Observable<boolean> {
+    return process.env.NODE_ENV === 'production'
+      ? super.canActivate(context)
+      : true;
+  }
+}
