+++ frontend/src/pages/order/components/index.ts
@@ -1,0 +1,2 @@
+export { default as EditOrder } from "./EditOrder";
+export { default as OrderList } from "./OrderList";
